Added role field in user creation and moved create user inside user management component
