+++ src/app/models/user.ts
@@ -2,4 +2,5 @@ export class User {
     username: string;
     password: string;
     confirmPassword:string;
+    role:number;
 }
