+++ src/app/user-management/create-user/create-user.component.html
@@ -3,13 +3,13 @@
   <div class="modal-card">
     <header class="modal-card-head">
       <p class="modal-card-title">Create User</p>
-      <button class="delete" aria-label="close" (click)="toggleModal(false)"></button>
+      <button class="delete" aria-label="close" (click)="toggleModal(false, f)"></button>
     </header>
     <section class="modal-card-body">
       <div class="card-content">
-        <form name="form" (ngSubmit)="f.form.valid && createUser()" #f="ngForm" novalidate>
+        <form name="form" (ngSubmit)="f.form.valid && createUser(f)" #f="ngForm" novalidate>
           <div class="columns">
-            <div class="column" [ngClass]="{ 'has-error': f.submitted && !username.valid }">
+            <div class="column">
               <label class="label">User Name</label>
               <p class="control is-fullwidth">
                 <input type="text" class="input" placeholder="username" name='username' [(ngModel)]="model.username" #username="ngModel"
@@ -19,7 +19,7 @@
             </div>
           </div>
           <div class="columns">
-            <div class="column" [ngClass]="{ 'has-error': f.submitted && !password.valid }">
+            <div class="column" >
               <label class="label">Password</label>
               <p class="control is-fullwidth">
                 <input type="password" class="input" placeholder="password" name='password' [(ngModel)]="model.password" #password="ngModel"
@@ -29,13 +29,24 @@
             </div>
           </div>
           <div class="columns">
-            <div class="column" [ngClass]="{ 'has-error': f.submitted && !confirmPassword.valid }">
+            <div class="column">
               <label class="label">Confirm Password</label>
               <p class="control is-fullwidth">
                 <input type="password" class="input" placeholder="confirm password" name='confirmPassword' [(ngModel)]="model.confirmPassword"
                   #confirmPassword="ngModel" validateEqual="password" required />
               </p>
               <small *ngIf="f.submitted && !confirmPassword.valid" class="help is-danger level-left">Password mismatch</small>
+            </div>
+          </div>
+          <div class="columns">
+            <div class="column" [ngClass]="{ 'has-error': f.submitted && !role.valid }">
+              <label class="label">Role</label>
+              <div class="select is-fullwidth">
+                <select id="day"   name='role' [(ngModel)]="model.role"  #role="ngModel">
+                  <option [value] ='1' selected >Admin</option>
+                  <option [value]='2'>User</option>
+                </select>
+              </div>
             </div>
           </div>
           <div class="columns">
