+++ src/app/user-management/create-user/create-user.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit, EventEmitter, Input, Output, Directive } from '@angular/core';
-import { UserService, AlertService } from '../services/index';
-import Utils from '../utils';
+import { UserService, AlertService } from '../../services/index';
+import Utils from '../../utils';
+import { User } from '../../models';
+import { NgForm } from '@angular/forms';
 
 @Component({
   selector: 'app-create-user',
@@ -8,16 +10,26 @@ import Utils from '../utils';
   styleUrls: ['./create-user.component.css']
 })
 export class CreateUserComponent implements OnInit {
-  model: any = {}
+  model: User;
 
   @Output() notify: EventEmitter<any> = new EventEmitter<any>();
 
   constructor(private userService: UserService,
     private alertService: AlertService) { }
 
-  ngOnInit() { }
+  ngOnInit() {
+    this.model = {
+      username: '',
+      password: '',
+      confirmPassword: '',
+      role: 1   // to set default value in role option
+    }
+  }
 
-  public toggleModal(isOpen: Boolean) {
+  public toggleModal(isOpen: Boolean, form: NgForm) {
+    if (form != null) {
+      form.resetForm({ role: 1 })
+    }
     let createUserModal = <HTMLDivElement>document.getElementById('user_modal');
     if (isOpen) {
       createUserModal.classList.add('is-active');
@@ -26,14 +38,17 @@ export class CreateUserComponent implements OnInit {
     createUserModal.classList.remove('is-active');
   }
 
-  public createUser() {
+  public createUser(form:NgForm) {
     let token = sessionStorage.getItem('token');
     this.userService.createUser(this.model, token).
       subscribe(
         data => {
           this.notify.emit();
-          this.toggleModal(false);
+          this.toggleModal(false, null);
           this.alertService.success(data.message);
+          if (form != null) {
+            form.resetForm({ role: 1 })
+          }
         },
         error => {
           if (error.status === 0) {
